refactor: provision scripts
package managers output redirection

+++ .assets/provision/install_kubectl.sh
@@ -27,7 +27,7 @@ SYS_ID=$(grep -oPm1 '^ID(_LIKE)?=.*\K(arch|fedora|debian|ubuntu|opensuse)' /etc/
 
 case $SYS_ID in
 arch)
-  pacman -Sy --needed --noconfirm kubectl >&2
+  pacman -Sy --needed --noconfirm kubectl >&2 2>/dev/null
   ;;
 fedora)
   [ -f /etc/yum.repos.d/kubernetes.repo ] || cat <<EOF | tee /etc/yum.repos.d/kubernetes.repo
@@ -38,23 +38,25 @@ enabled=1
 gpgcheck=1
 gpgkey=https://packages.cloud.google.com/yum/doc/yum-key.gpg https://packages.cloud.google.com/yum/doc/rpm-package-key.gpg
 EOF
-  dnf install -y kubectl >&2
+  dnf install -y kubectl >&2 2>/dev/null
   ;;
 debian | ubuntu)
   export DEBIAN_FRONTEND=noninteractive
-  apt-get update >&2 && apt-get install -y apt-transport-https ca-certificates curl >&2
+  apt-get update >&2 && apt-get install -y apt-transport-https ca-certificates curl >&2 2>/dev/null
   # download the Google Cloud public signing key
   curl -fsSLk -o /usr/share/keyrings/kubernetes-archive-keyring.gpg 'https://packages.cloud.google.com/apt/doc/apt-key.gpg'
   # add the Kubernetes apt repository
   [ -f /etc/apt/sources.list.d/kubernetes.list ] || echo "deb [signed-by=/usr/share/keyrings/kubernetes-archive-keyring.gpg] https://apt.kubernetes.io/ kubernetes-xenial main" | tee /etc/apt/sources.list.d/kubernetes.list >/dev/null
   # update apt package index with the new repository and install kubectl
-  apt-get update >&2 && apt-get install -y kubectl >&2
+  apt-get update >&2 && apt-get install -y kubectl >&2 2>/dev/null
   ;;
-*)
+esac
+
+if ! type $APP &>/dev/null; then
+  echo 'Installing from binary.' >&2
   while [[ ! -f kubectl ]]; do
     curl -LOsk "https://dl.k8s.io/release/$(curl -Lsk https://dl.k8s.io/release/stable.txt)/bin/linux/amd64/kubectl"
   done
   # install
   install -o root -g root -m 0755 kubectl /usr/bin/kubectl && rm -f kubectl
-  ;;
-esac
+fi
